requirements.txt
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
requests>=2.28.0
python-dateutil>=2.8.0
keyring>=23.0.0
Pillow>=9.0.0
bleach>=5.0.0
markdown>=3.4.0
pyspellchecker>=0.7.0
